instrumentation/python/metrics/gauge/requirements.txt
flask
psutil
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus
opentelemetry-instrumentation-flask